Download tables in different formats
Source: R/downloadTable.R

This module uses downloadHandler to allow downloads of tables in different formats (xlsx, rds, adn csv).

downloadTable(input, output, session, rTable, fileName = "tableFile")

downloadTableUI(id)

Arguments

input, output, session
Standard module parameters

rTable
A reactive table (for example a data.frame)

fileName
A prefix for the default name in the download handler.

id
The module identifier

Value

Currently, no reactive code is returned due to issues with downloadTable. See here

Examples

# NOT RUN {
shinyApp(
  fluidPage(downloadTableUI("downloadTable")),
  function(input, output, session) {
    callModule(downloadTable, "downloadTable", reactive({mtcars}), fileName = "mtcars")
  }
)
# }